requirements.txt
# Ikarus 3D Product Recommendation System - Python Dependencies
# Compatible with Python 3.12

# =============================================================================
# CORE WEB FRAMEWORK
# =============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# =============================================================================
# DATA PROCESSING & ANALYSIS
# =============================================================================
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4

# =============================================================================
# MACHINE LEARNING & AI
# =============================================================================
scikit-learn==1.3.2
torch==2.1.2
torchvision==0.16.2
transformers==4.35.2
sentence-transformers==2.2.2

# =============================================================================
# NLP AND TEXT PROCESSING
# =============================================================================
spacy==3.7.2
nltk==3.8.1

# =============================================================================
# COMPUTER VISION
# =============================================================================
opencv-python==4.8.1.78
Pillow==10.1.0

# =============================================================================
# VECTOR DATABASE & EMBEDDINGS
# =============================================================================
pinecone-client==2.2.4
faiss-cpu==1.7.4

# =============================================================================
# DATABASE & CACHING
# =============================================================================
sqlalchemy==2.0.23
redis==5.0.1

# =============================================================================
# HTTP & ASYNC
# =============================================================================
httpx==0.25.2
aiofiles==23.2.1

# =============================================================================
# ENVIRONMENT & CONFIGURATION
# =============================================================================
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# =============================================================================
# TESTING
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1

# =============================================================================
# DEVELOPMENT
# =============================================================================
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0

# =============================================================================
# PRODUCTION
# =============================================================================
gunicorn==21.2.0